DABOM-491 refactor: domain 이름 eventoutbox로 변경
