+++ .env.example
@@ -86,11 +86,10 @@ BATCH_DB_REDIS_RECONCILIATION_REDIS_CHUNK_SIZE=2000
 BATCH_DB_REDIS_RECONCILIATION_DB_FETCH_SIZE=4000
 
 # --- Batch Job Tuning: Usage Event Outbox ---
-BATCH_USAGE_EVENT_OUTBOX_BATCH_SIZE=100
-BATCH_USAGE_EVENT_OUTBOX_CONCURRENCY=8
-BATCH_USAGE_EVENT_OUTBOX_MAX_RETRY=5
-BATCH_USAGE_EVENT_OUTBOX_RETRY_INITIAL_DELAY=PT1M
-BATCH_USAGE_EVENT_OUTBOX_RETRY_MAX_DELAY=PT16M
-BATCH_USAGE_EVENT_OUTBOX_RETRY_ELIGIBILITY_DELAY=PT1M
-BATCH_USAGE_EVENT_OUTBOX_PUBLISH_TIMEOUT=PT10S
-BATCH_USAGE_EVENT_OUTBOX_TOPIC=notification
+BATCH_EVENT_OUTBOX_BATCH_SIZE=100
+BATCH_EVENT_OUTBOX_CONCURRENCY=8
+BATCH_EVENT_OUTBOX_MAX_RETRY=5
+BATCH_EVENT_OUTBOX_RETRY_INITIAL_DELAY=PT1M
+BATCH_EVENT_OUTBOX_RETRY_MAX_DELAY=PT16M
+BATCH_EVENT_OUTBOX_RETRY_ELIGIBILITY_DELAY=PT1M
+BATCH_EVENT_OUTBOX_PUBLISH_TIMEOUT=PT10S
