+++ src/main/java/com/template/worker/jobs/eventoutbox/model/EventOutboxRow.java
@@ -1,0 +1,3 @@
+package com.template.worker.jobs.eventoutbox.model;
+
+public record EventOutboxRow(long id, String eventId, String payloadJson, int retryCount) {}
