+++ src/main/java/com/template/worker/jobs/eventoutbox/support/EventOutboxJobConstants.java
@@ -1,0 +1,10 @@
+package com.template.worker.jobs.eventoutbox.support;
+
+public final class EventOutboxJobConstants {
+
+    public static final String JOB_NAME = "event-outbox-publish-job";
+    public static final String STEP_PUBLISH_USAGE_EVENT_OUTBOX = "publish-event-outbox-step";
+    public static final String KST_ZONE_ID_NAME = "Asia/Seoul";
+
+    private EventOutboxJobConstants() {}
+}
