+++ src/main/java/com/template/worker/jobs/eventoutbox/support/EventOutboxStatus.java
@@ -1,0 +1,7 @@
+package com.template.worker.jobs.eventoutbox.support;
+
+public enum EventOutboxStatus {
+    PUBLISH_PENDING,
+    FAILED,
+    SENT
+}
