+++ src/main/java/com/template/worker/jobs/eventoutbox/tasklet/EventOutboxPublishTasklet.java
@@ -1,0 +1,32 @@
+package com.template.worker.jobs.eventoutbox.tasklet;
+
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.stereotype.Component;
+
+import com.template.worker.jobs.eventoutbox.service.EventOutboxPublishService;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class EventOutboxPublishTasklet implements Tasklet {
+
+    private final EventOutboxPublishService publishService;
+
+    /** 한 번의 job execution에서 publish 가능한 outbox를 처리한다. */
+    @Override
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) {
+        int processedCount = publishService.publishPendingOutboxes();
+        if (processedCount > 0) {
+            log.info(
+                    "Processed usage event outbox rows in batch job. processedCount={}",
+                    processedCount);
+        }
+        return RepeatStatus.FINISHED;
+    }
+}
