--- src/main/java/com/template/worker/jobs/usageoutbox/model/UsageEventOutboxRow.java
@@ -1,3 +1,0 @@
-package com.template.worker.jobs.usageoutbox.model;
-
-public record UsageEventOutboxRow(long id, String eventId, String payloadJson, int retryCount) {}
