--- src/main/java/com/template/worker/jobs/usageoutbox/step/PublishUsageEventOutboxStepConfig.java
@@ -1,31 +1,0 @@
-package com.template.worker.jobs.usageoutbox.step;
-
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.step.builder.StepBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import com.template.worker.jobs.usageoutbox.support.UsageEventOutboxJobConstants;
-import com.template.worker.jobs.usageoutbox.tasklet.UsageEventOutboxPublishTasklet;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class PublishUsageEventOutboxStepConfig {
-
-    private final JobRepository jobRepository;
-    private final PlatformTransactionManager transactionManager;
-    private final UsageEventOutboxPublishTasklet tasklet;
-
-    /** usage outbox publish tasklet을 감싸는 단일 step을 구성한다. */
-    @Bean
-    public Step publishUsageEventOutboxStep() {
-        return new StepBuilder(
-                        UsageEventOutboxJobConstants.STEP_PUBLISH_USAGE_EVENT_OUTBOX, jobRepository)
-                .tasklet(tasklet, transactionManager)
-                .build();
-    }
-}
