--- src/main/java/com/template/worker/jobs/usageoutbox/support/UsageEventOutboxJobConstants.java
@@ -1,10 +1,0 @@
-package com.template.worker.jobs.usageoutbox.support;
-
-public final class UsageEventOutboxJobConstants {
-
-    public static final String JOB_NAME = "usage-event-outbox-publish-job";
-    public static final String STEP_PUBLISH_USAGE_EVENT_OUTBOX = "publish-usage-event-outbox-step";
-    public static final String KST_ZONE_ID_NAME = "Asia/Seoul";
-
-    private UsageEventOutboxJobConstants() {}
-}
