--- src/main/java/com/template/worker/jobs/usageoutbox/support/UsageEventOutboxStatus.java
@@ -1,7 +1,0 @@
-package com.template.worker.jobs.usageoutbox.support;
-
-public enum UsageEventOutboxStatus {
-    PUBLISH_PENDING,
-    FAILED,
-    SENT
-}
